backend/data-api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Scaffidi", "Zach Durham", "Ricardo Jimenez", "Andy Leeb"]

[tool.poetry.dependencies]
aiohttp = "^3.11.11"
asyncpg = "^0.30.0"
backoff = "^2.2.1"
bcrypt = "^4.3.0"
fastapi = "^0.115.12"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
pydantic = {extras = ["email"], version = "^2.11.1"}
python = "^3.13"
sendgrid = "^6.11.0"
uvicorn = "^0.30.6"
google-cloud-storage = "^2.18.2"
openai = "^1.69.0"
opensearch-py = "^2.7.1"
google-auth = "2.35.0"
pydantic-settings = "^2.8.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.2"
isort = "^5.13.2"
mypy = "^1.15.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.23.8"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api = "app.entrypoints:run_api"
vacuum = "app.commands.organizations.vacuum:run_command"
test_cleanup = "app.commands.organizations.cleanup_test_data:run_command"
reindex = "app.commands.organizations.reindex:run_command"

[tool.black]
line-length = 120
